pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "traffic-fines"
version = "0.2.0"
description = "Simulation framework for analyzing welfare effects of income-based vs flat traffic fines"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Max Ghenis" }
]
keywords = ["economics", "simulation", "traffic-fines", "optimal-taxation", "welfare"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.1",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "black>=24.0",
]
paper = [
    "jupyter-book[paper]>=2.1.0",
    "matplotlib>=3.8",
    "plotly>=5.18",
]
data = ["policyengine-us>=1.0"]

[project.urls]
Homepage = "https://github.com/MaxGhenis/traffic-fine-sim"
Repository = "https://github.com/MaxGhenis/traffic-fine-sim"

[tool.hatch.build.targets.wheel]
packages = ["src/traffic_fines"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=traffic_fines --cov-report=term-missing"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py311"]
